Split up IntoKeyword
