+++ tests/Components/IntoKeywordTest.php
@@ -4,40 +4,40 @@ declare(strict_types=1);
 
 namespace PhpMyAdmin\SqlParser\Tests\Components;
 
-use PhpMyAdmin\SqlParser\Components\IntoKeyword;
 use PhpMyAdmin\SqlParser\Parser;
+use PhpMyAdmin\SqlParser\Parsers\IntoKeywords;
 use PhpMyAdmin\SqlParser\Tests\TestCase;
 
 class IntoKeywordTest extends TestCase
 {
     public function testParse(): void
     {
-        $component = IntoKeyword::parse(new Parser(), $this->getTokensList('OUTFILE "/tmp/outfile.txt"'));
+        $component = IntoKeywords::parse(new Parser(), $this->getTokensList('OUTFILE "/tmp/outfile.txt"'));
         $this->assertEquals($component->type, 'OUTFILE');
         $this->assertEquals($component->dest, '/tmp/outfile.txt');
     }
 
     public function testBuild(): void
     {
-        $component = IntoKeyword::parse(new Parser(), $this->getTokensList('tbl(`col1`, `col2`)'));
+        $component = IntoKeywords::parse(new Parser(), $this->getTokensList('tbl(`col1`, `col2`)'));
         $this->assertEquals('tbl(`col1`, `col2`)', $component->build());
     }
 
     public function testBuildValues(): void
     {
-        $component = IntoKeyword::parse(new Parser(), $this->getTokensList('@a1, @a2, @a3'));
+        $component = IntoKeywords::parse(new Parser(), $this->getTokensList('@a1, @a2, @a3'));
         $this->assertEquals('@a1, @a2, @a3', $component->build());
     }
 
     public function testBuildOutfile(): void
     {
-        $component = IntoKeyword::parse(new Parser(), $this->getTokensList('OUTFILE "/tmp/outfile.txt"'));
+        $component = IntoKeywords::parse(new Parser(), $this->getTokensList('OUTFILE "/tmp/outfile.txt"'));
         $this->assertEquals('OUTFILE "/tmp/outfile.txt"', $component->build());
     }
 
     public function testParseErr1(): void
     {
-        $component = IntoKeyword::parse(new Parser(), $this->getTokensList('OUTFILE;'));
+        $component = IntoKeywords::parse(new Parser(), $this->getTokensList('OUTFILE;'));
         $this->assertEquals($component->type, 'OUTFILE');
     }
 }
